# perf(python): optimize `_datetime_to_pl_timestamp`