Similar to #9452 

On my MacBook Air M2

```python
In [1]: from datetime import datetime, timezone

In [2]: from polars.type_aliases import TimeUnit

In [3]: EPOCH_UTC = datetime(1970, 1, 1, tzinfo=timezone.utc)

In [4]: def _timestamp_in_seconds(dt: datetime) -> int:
   ...:     du = dt - EPOCH_UTC
   ...:     return du.days * 86400 + du.seconds
   ...:

In [5]: def _datetime_to_pl_timestamp(dt: datetime, time_unit: TimeUnit | None) -> int:
   ...:     """Convert a python datetime to a timestamp in nanoseconds."""
   ...:     dt = dt.replace(tzinfo=timezone.utc) if dt.tzinfo != timezone.utc else dt
   ...:     if time_unit == "ns":
   ...:         micros = dt.microsecond
   ...:         return 1_000 * (_timestamp_in_seconds(dt) * 1_000_000 + micros)
   ...:     elif time_unit == "us" or time_unit is None:
   ...:         micros = dt.microsecond
   ...:         return _timestamp_in_seconds(dt) * 1_000_000 + micros
   ...:     elif time_unit == "ms":
   ...:         millis = dt.microsecond // 1000
   ...:         return _timestamp_in_seconds(dt) * 1_000 + millis
   ...:     else:
   ...:         raise ValueError(
   ...:             f"time_unit must be one of {{'ns', 'us', 'ms'}}, got {time_unit}"
   ...:         )
   ...:

In [6]: def _datetime_to_pl_timestamp_main(dt: datetime, time_unit: TimeUnit | None) -> int:
   ...:     """Convert a python datetime to a timestamp in nanoseconds."""
   ...:     dt = dt.replace(tzinfo=timezone.utc)
   ...:     if time_unit == "ns":
   ...:         micros = dt.microsecond
   ...:         return 1_000 * (int(dt.replace(microsecond=0).timestamp() * 1_000_000) + micros)
   ...:     elif time_unit == "us" or time_unit is None:
   ...:         micros = dt.microsecond
   ...:         return int(dt.replace(microsecond=0).timestamp() * 1_000_000) + micros
   ...:     elif time_unit == "ms":
   ...:         millis = dt.microsecond // 1000
   ...:         return int(dt.replace(microsecond=0).timestamp() * 1_000) + millis
   ...:     else:
   ...:         raise ValueError(
   ...:             f"time_unit must be one of {{'ns', 'us', 'ms'}}, got {time_unit}"
   ...:         )
   ...:

In [7]: now = datetime.now()

In [8]: now_utc = now.replace(tzinfo=timezone.utc)

In [9]: %timeit _datetime_to_pl_timestamp_main(now, 'ns')
1.09 µs ± 12 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [10]: %timeit _datetime_to_pl_timestamp(now, 'ns')
655 ns ± 2.47 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [11]: %timeit _datetime_to_pl_timestamp_main(now_utc, 'ns')
1.11 µs ± 9.72 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [12]: %timeit _datetime_to_pl_timestamp(now_utc, 'ns')
221 ns ± 0.358 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```